스테이지에서 아이템이 등장하지 않는 버그 수정

+++ CastleHero-UnityClient/Assets/Scripts/Data/Model/StageEntity.cs
@@ -48,7 +48,7 @@ namespace RGLabs.Data.Model
         public int exp; 
 
         [DataField("Stage_Rwd_Item_Per")] 
-        public int itemPer;
+        public float itemPer;
 
         [DataField("Stage_Rwd_Item_ID")] 
         public int propItemId;
